Move unit tests under unit subdir
