+++ tests/unit/AvirWrapperTest.php
@@ -1,22 +1,24 @@
 <?php
 
 /**
- * Copyright (c) 2017 Victor Dubiniuk <dubiniuk@owncloud.com>
+ * Copyright (c) 2016 Victor Dubiniuk <victor.dubiniuk@gmail.com>
  * This file is licensed under the Affero General Public License version 3 or
  * later.
  * See the COPYING-README file.
  */
 
 
-namespace OCA\Files_antivirus\Tests;
+namespace OCA\Files_Antivirus\Tests\unit;
 
 use OC\Files\Filesystem;
-use OC\Files\View;
 use OC\Files\Storage\Storage;
 use OCA\Files_Antivirus\AvirWrapper;
+use OCA\Files_Antivirus\Tests\util\DummyClam;
 use Test\Util\User\Dummy;
 
-class ChunkUploadTest extends TestBase{
+
+class AvirWrapperTest extends TestBase {
+	
 	const UID = 'testo';
 	const PWD = 'test';
 
@@ -29,49 +31,49 @@ class ChunkUploadTest extends TestBase{
 		\OC_User::clearBackends();
 		\OC_User::useBackend(new Dummy());
 	}
-	
+
 	public function setUp() {
 		parent::setUp();
 		if (!\OC::$server->getUserManager()->get(self::UID)) {
 			\OC::$server->getUserManager()->createUser(self::UID, self::PWD);
 		}
-		
-		$this->scannerFactory = $this->getMockBuilder(\OCA\Files_Antivirus\ScannerFactory::class)
-			->setConstructorArgs([
-				new Mock\Config($this->container->query('CoreConfig')),
-				$this->container->query('Logger')
-			])
-			->setMethods(['getScanner'])
-			->getMock();
+
+		$this->scannerFactory = new Mock\ScannerFactory(
+			new Mock\Config($this->container->query('CoreConfig')),
+			$this->container->query('Logger')
+		);
 
 		if (!$this->isWrapperRegistered) {
 			Filesystem::addStorageWrapper(
-				'oc_avir_test_chunk',
+				'oc_avir_test',
 				[$this, 'wrapperCallback'],
-				3
+				2
 			);
 			$this->isWrapperRegistered = true;
 		}
+
 		\OC::$server->getUserSession()->login(self::UID, self::PWD);
 		\OC::$server->getSession()->set('user_id', self::UID);
 		\OC::$server->getUserFolder(self::UID);
 	}
 
-	public function testSkipIndividualChunks(){
-		$this->scannerFactory->expects($this->never())
-			->method('getScanner');
-
-		$path = '/' . self::UID . '/uploads';
-		$rootView = new View();
-		if (!$rootView->file_exists($path)) {
-			$rootView->mkdir($path);
-		}
-		$view = new View($path);
-		$fd = $view->fopen('/chunk', 'w+');
-		@fwrite($fd, 'abcdead');
-		@fclose($fd);
+	/**
+	 * @expectedException \OCP\Files\InvalidContentException
+	 */
+	public function testInfected(){
+		$fd = Filesystem::fopen('killing bee', 'w+');
+		@fwrite($fd, 'it ' . DummyClam::TEST_SIGNATURE);
 	}
-	
+
+	/**
+	 * @expectedException \OCP\Files\InvalidContentException
+	 */
+	public function testBigInfected(){
+		$fd = Filesystem::fopen('killing whale', 'w+');
+		@fwrite($fd, str_repeat('0', DummyClam::TEST_STREAM_SIZE-2) . DummyClam::TEST_SIGNATURE );
+		@fwrite($fd, DummyClam::TEST_SIGNATURE);
+	}
+
 	public function wrapperCallback($mountPoint, $storage){
 		/**
 		 * @var Storage $storage
