+++ tests/unit/bootstrap.php
@@ -7,7 +7,7 @@ if (!defined('PHPUNIT_RUN')) {
 	define('PHPUNIT_RUN', 1);
 }
 
-require_once __DIR__.'/../../../lib/base.php';
+require_once __DIR__.'/../../../../lib/base.php';
 
 \OC::$composerAutoloader->addPsr4('Test\\', OC::$SERVERROOT . '/tests/lib/', true);
 \OC::$composerAutoloader->addPsr4('Tests\\', OC::$SERVERROOT . '/tests/', true);
