+++ tests/util/avir.sh
